refactor: move metrics tracker into plugin reporting
Moved createMetricsTracker from src/shared/utils/metrics.js (and its src/shared/metrics-utils shim) into src/plugin/src/reporting/metrics.js because the tracker only serves the plugin and keeping it in shared blurred package boundaries. Updated shared utils barrel exports and relocated the metrics unit test into src/plugin/tests so imports and coverage stay aligned.
