+++ src/plugin/tests/metrics-utils.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../metrics-utils.js";
+import { createMetricsTracker } from "../src/reporting/metrics.js";
 
 function getCacheKeys(tracker, cacheName = "example") {
     tracker.recordCacheHit(cacheName);
